requirements.txt
click==7.1.2
requests==2.25.0
inject==4.3.1
git+https://github.com/anodot/sdc-client
pyyaml==5.3.1
flask==1.1.2
